Commit for December 5th meeting.
Note for self and Professor Higham: The entire
NBA project is now included in
Olympics_Project_Write_Up.qmd, so
NBA_Wrangling.qmd can be ignored, and we only
need to look at the write-up in our meeting.
(The specified write-up edits for the Olympics
project were also completed in here).